modules/person_service/requirements.txt
SQLAlchemy==1.3.19
grpcio==1.39.0
grpcio-tools==1.39.0
psycopg2-binary==2.8.5